+++ R/review.R
@@ -1290,6 +1290,10 @@ REV_report_changes <- function(h0, h1, verbose = FALSE) {
 
 #' Early error feedback function for the optional review parameter
 #'
+#' @param afmm
+#' 
+#' Pass-through of the server afmm parameter.
+#' 
 #' @param datasets `[list(data.frame)]`
 #'
 #' Available datasets for review.
@@ -1307,7 +1311,7 @@ REV_report_changes <- function(h0, h1, verbose = FALSE) {
 #' the configuration of the review parameter will be placed here.
 #' 
 #' @export
-check_review_parameter <- function(datasets, dataset_names, review, err) {
+check_review_parameter <- function(afmm, datasets, dataset_names, review, err) {
   if (is.null(review)) return(NULL)
   ok <- CM$assert(
     container = err,
@@ -1443,4 +1447,48 @@ check_review_parameter <- function(datasets, dataset_names, review, err) {
       )
     }
   }
+  if (!ok) return(NULL)
+ 
+  # https://en.wikipedia.org/wiki/Filename#Problematic_characters 
+  problematic_chars <- c("/", "\\\\", "?", "%", "*", ":", "|", '"', "<", ">", ".", ",", ";", "=")
+  problematic_chars_regexp <- paste("([", paste(problematic_chars, collapse = ""), "])")
+  
+  report_problematic_names <- function(message_template, v) {
+    res <- character(0)
+    for (e in v){
+      matches <- character(0)
+      for (pc in problematic_chars) if (grepl(pc, e, fixed = TRUE)) {
+        single_char <- substr(pc, 1, 1) # deals with backslash
+        matches <- c(matches, single_char)
+      }
+      if (length(matches)) {
+        res <- c(res, sprintf(message_template, e, paste(sprintf("'<b>%s</b>'", matches), collapse = ", ")))
+      }
+    }
+    res <- paste(res, collapse = "<br>")
+    return(res)
+  }
+  
+  dataset_list_names <- names(afmm[["data"]])
+  CM$assert(
+    container = err,
+    cond = !any(grepl(problematic_chars_regexp, dataset_list_names)),
+    msg = report_problematic_names(
+      paste('The dataset list name "<b>%s</b>" contains characters (%s) incompatible with the review functionality.',
+            "That string would be used as part of review-related folder names and those characters could cause problems",
+            '<a href="https://en.wikipedia.org/wiki/Filename#Problematic_characters" target="_blank">(details)<a>.',
+            "Please, exclude them."),
+      dataset_list_names)
+  ) 
+  dataset_names <- names(review[["datasets"]])
+  CM$assert(
+    container = err,
+    cond = !any(grepl(problematic_chars_regexp, dataset_names)),
+    msg = report_problematic_names(
+      paste('The dataset name "<b>%s</b>" contains characters (%s) incompatible with the review functionality.',
+            "That string would be used as part of review-related file names and those characters could cause problems",
+            '<a href="https://en.wikipedia.org/wiki/Filename#Problematic_characters" target="_blank">(details)<a>.',
+            "Please, exclude them."),
+      dataset_names)
+  )
 }